+++ src/views/admin/default/index.jsx
@@ -147,8 +147,8 @@ export default function MainDashboard() {
           startContent={
             <IconBox
               w='56px' h='56px'
-              bg='linear-gradient(90deg, #4481EB 0%, #04BEFE 100%)'
-              icon={<Icon w='28px' h='28px' as={MdPersonAdd} color='white' />}
+              bg={boxBg}
+              icon={<Icon w='32px' h='32px' as={MdPersonAdd} color={brandColor} />}
             />
           }
           name='신규 방문'